[quickview] Test the PDF file is displayed when multiple files are selected
Check-select an image and a PDF file, show the files in Quick View,
and down-arrow to the PDF file.

Check the PDF file is displayed in Quick View (tests a bug where the
PDF file was not being displayed, which was fixed in CL:2007795).

Bug: 1038799
Change-Id: I5e6eda55c64458f7fd82a343597906399ba5feda
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2010602
Commit-Queue: Noel Gordon <noel@chromium.org>
Reviewed-by: Noel Gordon <noel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733506}