media/gpu/v4l2svda: fix GL image size
The size of the GL image is supposed to be the visible size, while the
buffer itself shall be imported with its own coded size and stride.

Before crrev.com/c/1982499, we did not pass any size information with
the buffer, so as a workaround the size was set to the coded size and
the buffer imported with its coded size into the GPU or DC, which
resulted in visible green lines on the right or lower end of the image
when coded_size != visible_size.

Now that we are importing the buffer using a NativePixmapHandle, we can
set the GL image to be the correct visible size.

BUG=b:141579960
TEST=240p h.264 video plays in Youtube web without visible green line on
Kukui.
TEST=video playback working on Kevin.

Change-Id: I870892c41197ed210c7b7455975669a3473891f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2010243
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Auto-Submit: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733513}