Reformat all gn files in /chromeos/services/machine_learning.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=amoylan@chromium.org

Bug: 1041419
Change-Id: I74c66ddca81f15a70cbd52643a9ee6817181c0dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1998577
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Commit-Queue: Andrew Moylan <amoylan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733527}