Label responds to theme change
Previously, Label would only update colors in response to theme changes
if the text_style was STYLE_PRIMARY.

CustomTabBarView now uses CONTEXT_LABEL for the title and the origin.
As the origin has STYLE_SECONDARY, this causes the origin to be faded
relative to the title.

Screen shots before and after:
https://docs.google.com/document/d/1YRuzw0Bh9IyjmbtwBa0NpwNvu7ttbMM8mdRF0K3YWzc/edit?usp=sharing

Bug: 1033298
Change-Id: Iff4dbc7dfe611cd4c3a38da0b39fc790dae29bcc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2010757
Commit-Queue: Eric Willigers <ericwilligers@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733552}