[AutofillAssistant] Make more resilient by checking flags first.
After this change it is possible to completely disable all code
execution / changes by disabling the AutofillAssistantChromeEntry
feature.
This also goes for the AutofillAssistant code, waiting for initial
navigation.

This CL also removes the temporary workaround that was setting the
fallback URL in code if it was missing in the intent.

Bug: b/145195894
Change-Id: Ic35b799603c633a310171d9f40574e3ff8d79b7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2007696
Commit-Queue: Sandro Maggi <sandromaggi@google.com>
Reviewed-by: Mathias Carlen <mcarlen@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733598}