Reformat all gn files in /chromecast/
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by me.

Bug: 1041419
Change-Id: I872f425930fdc439c8b807abd87199cea0c058c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2012601
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Luke Halliwell (slow) <halliwell@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733611}