[Autofill Assistant] Allow partial overlays in collect user data action
At the moment the touchable elements from the FocusElement action are
only applied to the overlay if a Prompt or Show Form action follows it.
In this CL we make so that CollectUserData also sets the partial overlay
the same way as Prompt and Show Form do.

Bug: b/146483444
Change-Id: Ica6bcb0928ee6ed35005ace3831614ccd888d3c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2002818
Commit-Queue: Luca Hunkeler <hluca@google.com>
Reviewed-by: Mathias Carlen <mcarlen@chromium.org>
Reviewed-by: Stephane Zermatten <szermatt@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733624}