Reland "Fix flaky test ServerAddressConvertsToSameLocalAddress"
This is a reland of commit b3873b4d03efae2aa1e40f73657dd9f3b88e73dc.

The CL fixes a race condition in production code of autofill profile
sync bridge by modifying merge logic of the datatype. Previously, the
logic for dealing with duplicate entities was to always prefer the
remote one and silently delete the local one. This could end up with
two clients having a different entry each and the sync server having
both duplicate entries. The new behavior, introduced in this CL is to
keep the profile with greater GUID, lexicographically.

This change should also remove flakiness of one integration test (that
uncovered the actual issue), so the CL reenables the test.

The original CL caused flakiness in another integration test since GUIDs
in that test are randomly attributed. In this reland, we fix this by
making the expectation on the number of sync deletions depend on which
one of the test GUIDs is bigger.

Bug: 917498
Change-Id: I13994ac5bef8fcae41d37350e08ee80ab90f0916
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2002571
Auto-Submit: Jan Krcal <jkrcal@chromium.org>
Commit-Queue: Jan Krcal <jkrcal@chromium.org>
Commit-Queue: Maxim Kolosovskiy <kolos@chromium.org>
Reviewed-by: Maxim Kolosovskiy <kolos@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733636}