[CompositeAfterPaint] Record foreign layers for remote frames
An out-of-process (remote) frame has a SurfaceLayer in the parent
render process. This surface layer is used by viz to combine the child
and parent surfaces together. This patch ensures we pass the
SurfaceLayer to cc. This was done in
CLM::UpdateGraphicsLayerConfiguration before CompositeAfterPaint.

Bug: 959936, 1041496, 1043320, 1043327
Change-Id: Iaa91051a6b8d1b9fb973596888476e4714d62a7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2011162
Commit-Queue: Philip Rogers <pdr@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733649}