Reformat all gn files in /extensions/
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by me.

Bug: 1041419
Change-Id: I0ca2925d8e45744cbbdbb89aff63fb97fa5cf4d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2012600
Auto-Submit: Nico Weber <thakis@chromium.org>
Commit-Queue: Ken Rockot <rockot@google.com>
Reviewed-by: Ken Rockot <rockot@google.com>
Cr-Commit-Position: refs/heads/master@{#733654}