Do not use original URL/method for error page navigation params
The WebNavigationParams, which are eventually passed into
DocumentLoader, contain a URL and method that are set to the original
URL/method of the request. When a page loads successfully,
FillNavigationParamsResponseAndBodyLoader() populates the |redirects|
field on the params, and when DocumentLoader starts loading, it uses
|redirects| to replay the redirect chain and update the URL and
method. However, when an error page is loaded (via
RenderFrameImpl::CommitFailedNavigation),
FillNavigationParamsRequest() populates the params with the original
URL and method, but FillNavigationParamsResponseAndBodyLoader() is not
called to populate the |redirects| field. (I *think* this is correct
behavior, as it would be odd to have |redirects| populated on the
DocumentLoader for chrome-error://chromewebdata.) This means that the
URL and method are never updated to follow the redirect chain as
DocumentLoader is loading the error page. The end result is that
reloading the error page will use the original method of the request,
because that is what DocumentLoader believes the method is (and what
it reports to the browser process which is what ends up getting
recorded on the FrameNavigationEntry).

This CL fixes this by filling the WebNavigationParams with the final
method in CommitFailedNavigation().

Bug: 1041597
Change-Id: Ib6a6616a7bb79b5dd186d5e05312fde992d4f6dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2006473
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Commit-Queue: Emily Stark <estark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733681}