Make sure KeySystems is up to date before selecting key system
UpdateIfNeeded() was only called in GetInstace(), but
KeySystemConfigSelector caches the KeySystems pointer.  Calls to
SelectConfig() that happened after installing a new key system still saw
the initial list of key systems.

To fix this, we expose UpdateIfNeeded() in the KeySystems interface so
that KeySystemConfigSelector can call it in SelectConfig() to make sure
the information on supported key systems is up to date.

Bug: 1033932
Change-Id: I7bdbe7ead15907fa6605a445f6bf9d7842f3c1ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1967188
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Commit-Queue: Wojciech Dzierżanowski <wdzierzanowski@opera.com>
Auto-Submit: Wojciech Dzierżanowski <wdzierzanowski@opera.com>
Cr-Commit-Position: refs/heads/master@{#733688}