Hide Managed Guest Session user pod in lock screen
This CL updates the lock screen UI to not show public account user pods
when in the lock screen.

This is a follow-up of crrev.com/c/1958049. Managed guest sessions are
now possibly lock-able, but the lock screen UI currently reuses the
public account pod for the login screen. |ScreenLocker| now has a new
|GetUsersToShow()| method which returns the list of user pods which
should be shown. |LockContentsView| is updated to be able handle when
there are no user pods to be shown.

No user pod UI is intended to be shown for Managed Guest Sessions, the
unlocking of the session is handled by a login screen extension.

Bug: 1032065
Change-Id: I68f91fbdd5d576851a4929328548e16688ef2168
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2003326
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Jit Yao Yap <jityao@google.com>
Cr-Commit-Position: refs/heads/master@{#733689}