Attempt to fix race in unittests
What: DiskCacheBackendTest.SimpleCacheCalculateSizeOfEntriesBetween
started to flake after landing a change to enable prioritized
SimpleCache queues by default. I couldn't reproduce locally but my
first guess is that there is a race between creating the entries on
the worker thread and the two calls to calculate cache size on the
main thread. I noticed that the "FlushQueueForTest()" method that the
test calls doesn't do anything for simple caches. So I've updated it
to flush the simple cache's worker pool as well. Let's see if that
fixes things.

Bug: 1042441
Change-Id: I0a36da8f3e397f04c5241ec398671549096e7348
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2012582
Reviewed-by: Maksim Orlovich <morlovich@chromium.org>
Commit-Queue: Josh Karlin <jkarlin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733691}