webauthn: allow use in extensions.
Extensions don't have a domain name, which begs the question of what the
RP ID for an extension should be. This change defines the RP ID for
extensions to be their origin, i.e. “http(s)” is now special in that the
origin gets its scheme and port removed in order to form an RP ID. For
other origins that doesn't happen and thus the RP ID is just the origin
string.

Since extensions are not a concept in content/, allow the embedder to
handle this mapping and add a mapping for chrome-extension:// in the
Chrome embedder code.

Change-Id: Ie6453469404a81f5589843177d935c07eac8b59b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1979591
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Commit-Queue: Adam Langley <agl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733692}