mac: remove mac-views-task-manager flag
This has never been on by default and it doesn't look like it will be
any time soon. We'll stick with the Cocoa task manager for the
foreseeable future, so remove the flag and stop compiling the Views
task manager or its tests for Mac.

Bug: 836957
Change-Id: I0ed80f4e57c8bea11b46e9992c109e446b446fda
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2012881
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733696}