CrOS Shelf: Create new drag handle layer for the hotseat transition animation
Since an animating background layer is already used to visually
animate the position of the ShelfWidget's |opaque_background_|,
this change also creates an animating layer for the drag handle.
This animating drag handle is a child of the animating background
layer. This makes the animating drag handle follow the already
existing animation for the animating background layer.

Bug: 1040303
Change-Id: I452fba942f88e7194690d8937c22b10b7907a7d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2001852
Reviewed-by: Manu Cornet <manucornet@chromium.org>
Commit-Queue: Matthew Mourgos <mmourgos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733703}