Reformat all gn files in /third_party/proguard.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=agrieve@chromium.org

Bug: 1041419
Change-Id: Ia44520196c6007b0c9c8802d9d8e122192d2252b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997824
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733718}