Pass signon_realm as std::string instead of GURL
Use signon_realm as std::string instead of GURL because
the signon_realm in Android is a string in password_specifics.
To make it compatible for all platforms and easier to maintain,
use a std::string all around. The only method that isn't touched
here is RemoveRowsByUrlAndTimea as it uses a filter that needs a
GURL.

Bug: 1042117
Change-Id: I7c63981d4dd608ea22da8906a2d4144c08e8787c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2001346
Commit-Queue: Bettina Dea <bdea@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733721}