[identity] Implement GaiaRemoteConsentFlow
This CL implements display of the remote consent dialog for the new
consent flow. As before, the consent page is opened in a separate window
within the IdentityApiUiApp.

The new dialog, however, doesn't wait until the page is redirected or
the page title changed. The consent page passes the consent result via
private JS api that is yet to be implemented in the next CL. For now,
the browser never gets the result back.

After the consent result is get, the browser should send a request to
IssueToken endpoint. This part hasn't been implemented yet. (see
NOTIMPLEMENTED() in
IdentityGetAuthTokenFunction::OnGaiaRemoteConsentFlowCompleted()).

Bug: 1026237
Change-Id: I105f4036a4b2ae85eed4b29d810b9fe3bea2c6a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997303
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733738}