[iOS] Add PasswordInfobarModalInteractionHandler
This interaction handler is a subclass of InfobarModalInteractionHandler
that exposes functionality handling the dispatch callbacks that can be
sent for the InfobarPasswordModalDelegate.

Additionally, this CL adds a callback installer subclass that handles
setting up the callbacks to handle these dispatched comments and forward
them to the interaction handler.

Bug: 1030357
Change-Id: I07e5b872ebd1c8998333a1ad9369529942bf5fd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2001525
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733742}