Ozone/DRM: Eliminate tracking of owned planes in HardwareDisplayController
Don't keep track of owned planes directly in HardwareDisplayController.
We already know the set of owned CRTCs and we can ask the plane manager
to find the planes corresponding to those CRTCs. By removing
|owned_hardware_planes_| from HardwareDisplayController, we reduce
the amount of state that has to be protected by a lock if we wish to
perform pageflip testing on another thread (see the bug for more
context).

TEST=Built and ran on eve and chell. Connected and disconnected an
external monitor. There is also plenty of existing unit tests that cover
this code.

Bug: 1034559
Change-Id: I16fccc549c272c8659d392996a163877ac7bd7ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2003561
Commit-Queue: Saman Sami <samans@chromium.org>
Reviewed-by: Daniel Nicoara <dnicoara@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733746}