Add code to generate PROMPT_FOR_SCANNING verdicts
This CL adds the code needed to generate PROMPT_FOR_SCANNING danger
types. This danger type is used when an Advanced Protection user
downloads a file that is marked as UNCOMMON. It causes the user to be
prompted to send this file to Google for deep scanning. The generation
of these verdicts is guarded by a Finch flag as well as the user's
Advanced Protection enrollment status.

It also removes the old Finch flag related to Safe Browsing performing
additional download scanning for Advanced Protection users, to
minimize confusion.

Bug: 1020418
Change-Id: I506c1ff5fd3d923f8817f4db16563996d937a418
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2007984
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Xing Liu <xingliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733747}