Reparent the entire TabModel when the theme is swapped
As a follow-up to reparenting the current Tab when the theme is changed,
this CL reparents the entire TabModel. When the theme is swapped, first
the foreground tab is detached, then all other tabs are detached in tab
index order. When the app restarts, the tabs are reattched in reverse
tab index order (attaching the last, second to last), added as the
first tab with the exception of the foreground tab, which is left out.
After the tabs are stacked, restoring the original ordering, the
original foreground tab is added at the correct index.

Bug: 944302,1031332
Change-Id: I7eb1465157500e7f1f6c1da79240c96f38c9b6da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1974992
Commit-Queue: Brandon Wylie <wylieb@chromium.org>
Reviewed-by: Yusuf Ozuysal <yusufo@chromium.org>
Reviewed-by: Filip Gorski <fgorski@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733748}