Implement applying the extensions allowlist to both CORB and OOR-CORS.
Before this CL, the extensions allowlist (see [1]) applied only to CORB.
This meant that requests from non-allowlisted content scripts (to a
destination the extension has permission to) would be allowed by CORS,
but may be blocked by CORB.  Not only was this inconsistency bad from
hygiene/simplicity perspective, but it also was hurting content script
ability to issue requests that are allowed by CORS (ones where the
server would reply with Access-Control-Allow-Origins if given an Origin
header with the origin of the webpage where content scripts got
injected).

This CL implements applying the allowilist to both CORB and OOR-CORS.
The CL adds test coverage for this new behavior, but for now the new
behavior is behind a disabled-by-default base::Feature.

For more information, please see the design doc at
https://docs.google.com/document/d/15QPPRwnWl5YboQ6foTM65IfFW0KHRPqHHk-IoGN04-Y

Bug: 920638
Change-Id: I37325535d3633e5d44fd627b844032fbac550a74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1993759
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733768}