Split WebAppTabHelper into public interface and impl
This CL moves the implementation of WebAppTabHelper out of
web_applications/components down into a new web_applications:common
source set.
The publicly accessible methods of the class are retained in
WebAppTabHelperBase.

This is in preparation for migrating some of its dependencies out of
web_applications/components. Many of the classes in components should
be one level up as they're not isolated components or part of the
public interface of web_applications.

Bug: 1035242
Change-Id: I9d660e3450f8d0bb7ff7a5346346d117cabd6259
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1972741
Auto-Submit: Alan Cutter <alancutter@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Alexey Baskakov <loyso@chromium.org>
Commit-Queue: Alan Cutter <alancutter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733840}