WebApps: Implement unregistering file handlers on Linux.
This will allow us to disable the file handlers for an app on Linux
without uninstalling it. This is necessary preparation for the origin
trial, as we need to be able to disable/enable the file handlers as
the validity of the app's origin trial changes.

Bug: 1028448
Change-Id: I63eb6582852ff75d3d88181ede07f0cbd010df2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2010234
Commit-Queue: Jay Harris <harrisjay@chromium.org>
Reviewed-by: Alexey Baskakov <loyso@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733859}