Stop using chromeos::FakeSessionManagerClient in ArcSessionImpl tests
It's no longer necessary as we've introduced our own abstraction
layer, arc::ArcClientAdapter. This CL injects a fake ArcClientAdapter
implementation to the test fixture to simplify the test. That allows
us to remove all usage of chromeos:: and login_manager:: from the
test, and the test no longer depends on the internals of the ARC
container.

BUG=b:143175953
TEST=try

Change-Id: Ia39f492a94b5698794df423b181ff935b944eec4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2008284
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Yusuke Sato <yusukes@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733888}