[TTS] Fix the boundary for an enum histogram.
The histogram Search.ContextualSearchQuickActions.Category is
being written with an incorrect upper-bound that is too low.

The values written are in the range 0-5 which matches the enum in
enums.xml.  The problem is that the supplied upper-bound for the value
is being incorrectly supplied as being 3 rather than 6. This has the
effect of lumping all the high values into the third bucket.

Here's what tipped me off: when I viewed the histogram data I could see
there were only 3 buckets and the last had a wonky max value.
Pic here: https://screenshot.googleplex.com/BiZRJPYwSAi.png

This just fixes the supplied upper-bound to be correct, and extends the
histogram out for a year.

Change-Id: I11ba80ac0429468812b92817d85493091e987b11
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2003574
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Commit-Queue: Donn Denman <donnd@chromium.org>
Auto-Submit: Donn Denman <donnd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733897}