SSLErrorHandler: Configure triggering of extension events from client
SSLErrorHandler is on course to being componentized for sharing with
WebLayer. It currently internally triggers extension events when
security interstitials are shown. As this is //chrome-level behavior
that cannot be componentized, this CL gives SSLErrorHandler an API
via which the client can set a callback that is triggered on showing
of security interstitials.

Bug: 1030692
Change-Id: Ic3846d62c598af55ae64d880b229b37f1a12390c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2010943
Reviewed-by: Carlos IL <carlosil@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733912}