[Kiosk] Refactor out ArcKioskSplashScreen
Here we remove redundant ArcKioskSplashScreen, which had the same
functionality as AppLaunchSplashScreen, which ArcKioskController is
now using.

Also, now the splash screen of the arc kiosk properly displays the data
about the app. Also, this CL fixes a bug about not properly caching
the title and the icon of the arc kiosk app.

Bug: 1043663
Change-Id: I81c6a30021097041bd55a3a7182943fe9f93f3e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2007824
Commit-Queue: Anatoliy Potapchuk <apotapchuk@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Reviewed-by: Sergey Poromov <poromov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#733960}